graph: Move LDAP groups related code to a separate file
